requirements.txt
absl-py==2.4.0
addict==2.4.0
asgiref==3.11.1
attrs==26.1.0
basicsr==1.4.2
certifi==2026.2.25
charset-normalizer==3.4.6
cloudinary==1.44.1
coloredlogs==15.0.1
contourpy==1.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cycler==0.12.1
dj-database-url==3.1.2
Django==5.2.12
django-cloudinary-storage==0.3.0
django-jazzmin==3.0.4
djangorestframework==3.17.1
facexlib==0.3.0
filelock==3.25.2
filterpy==1.4.5
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.80.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-loader==0.5
llvmlite==0.46.0
lmdb==2.2.0
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
mysqlclient==2.2.8
networkx==3.4.2
numba==0.64.0
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
onnxruntime-gpu==1.23.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
protobuf==7.34.1
PyMatting==1.1.15
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
realesrgan==0.3.0
referencing==0.37.0
rembg==2.0.69
requests==2.33.0
rpds-py==0.30.0
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sqlparse==0.5.5
stripe==15.0.1
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
tifffile==2025.5.10
tomli==2.4.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
yapf==0.43.0
